api/requirements.txt
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
cx-Oracle==8.0.1
Flask==1.1.2
Flask-APScheduler==1.11.0
Flask-SQLAlchemy==2.4.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
paramiko==2.7.2
pkg-resources==0.0.0
pycparser==2.20
PyJWT==1.7.1
PyNaCl==1.4.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
requests==2.25.0
six==1.15.0
SQLAlchemy==1.3.20
sshtunnel==0.3.1
twilio==6.49.0
tzlocal==2.1
urllib3==1.26.2
Werkzeug==1.0.1
